--- comment-service/src/main/java/com/microservice/commentservice/CommentServiceApplication.java
@@ -1,20 +1,0 @@
-package com.microservice.commentservice;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.context.annotation.Bean;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class CommentServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CommentServiceApplication.class, args);
-	}
-	@Bean
-	public ModelMapper modelMapper() {
-		return new ModelMapper();
-	}
-}
