--- comment-service/src/main/java/com/microservice/commentservice/api/form/CommentContent.java
@@ -1,11 +1,0 @@
-package com.microservice.commentservice.api.form;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class CommentContent {
-   @NotNull
-    private String content;
-}
