--- comment-service/src/main/java/com/microservice/commentservice/api/form/CommentForm.java
@@ -1,11 +1,0 @@
-package com.microservice.commentservice.api.form;
-import lombok.Data;
-@Data
-public class CommentForm {
-    private Long id;
-    private String content;
-    private String createdDate;
-    private String updatedDate;
-    private Long userId;
-    private String img;
-}
