+++ comment-service/src/main/java/com/microservice/historyservice/api/form/CommentContent.java
@@ -1,0 +1,11 @@
+package com.microservice.historyservice.api.form;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class CommentContent {
+   @NotNull
+    private String content;
+}
