+++ comment-service/src/main/java/com/microservice/historyservice/api/form/CommentForm.java
@@ -1,0 +1,11 @@
+package com.microservice.historyservice.api.form;
+import lombok.Data;
+@Data
+public class CommentForm {
+    private Long id;
+    private String content;
+    private String createdDate;
+    private String updatedDate;
+    private Long userId;
+    private String img;
+}
