+++ history-service/pom.xml
@@ -9,15 +9,21 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.microservice</groupId>
-	<artifactId>commentservice</artifactId>
+	<artifactId>historyservice</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>comment-service</name>
-	<description>comment-service</description>
+	<name>history-service</name>
+	<description>history-service</description>
 	<properties>
 		<java.version>1.8</java.version>
 		<spring-cloud.version>2021.0.1</spring-cloud.version>
 	</properties>
 	<dependencies>
+		<dependency>
+			<groupId>org.modelmapper</groupId>
+			<artifactId>modelmapper</artifactId>
+			<version>2.3.0</version>
+		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
