+++ history-service/src/main/java/com/microservice/historyservice/HistoryServiceApplication.java
@@ -1,0 +1,15 @@
+package com.microservice.historyservice;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class HistoryServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HistoryServiceApplication.class, args);
+	}
+}
