+++ history-service/src/main/java/com/microservice/historyservice/api/form/HistoryForm.java
@@ -1,0 +1,10 @@
+package com.microservice.historyservice.api.form;
+
+import lombok.Data;
+
+@Data
+public class HistoryForm {
+    private Long userId;
+    private String actions;
+    private String createdDate;
+}
