+++ history-service/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+# ===============================
+# DATABASE
+# ===============================
+spring.datasource.url=jdbc:mysql://localhost:3306/pa_task_management
+spring.datasource.username=root
+spring.datasource.password=123456
+
+
+# ===============================
+# JPA / HIBERNATE
+# ===============================
+spring.user.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.jpa.show-sql=false
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
