+++ history-service/src/main/resources/application.yml
@@ -1,0 +1,14 @@
+server:
+  port: 9006
+
+spring:
+  application:
+    name: HISTORY-SERVICE
+  cloud:
+    config:
+      enabled: true
+      uri: http://localhost:9296
+  config:
+    import: "optional:configserver:"
+  zipkin:
+    base-url: http://127.0.0.1:9411/
