+++ history-service/src/test/java/com/example/commentservice/CommentServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.commentservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommentServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
